moves some files around for gemspec
