+++ lib/brewer/adaptibrew.rb
