+++ lib/brewer/helpers.rb
