+++ tests/ts_all.rb
@@ -3,7 +3,7 @@ require 'simplecov'
 SimpleCov.command_name 'Unit Tests'
 SimpleCov.start
 
-require_relative '../lib/helpers'
+require_relative '../lib/brewer/helpers'
 include Helpers
 
 # Test Framework
